`helm diff` supports new releases (#72)
* Bump kubernetes version 1.9.6 -> 1.10.4

* Switch Dockerfile `ENV` to `ARG`. `ARG` lets users provide their own
value at build-time. The environment variable is no longer set at
runtime, but I don't believe it is used.

* `helm diff` supports new releases

helm diff v2.9.0+2 adds the `--allow-unreleased` flag which shows the
full helm release as a diff if the release is brand new. Without this
flag, helm diff errors. Currently, the workaround is to not provide the
`helm_diff: true` parameter during the first `put` and to only add it
later.