Allow setting values as string (#78)
* reactoring in override_values processing

* add support for `type: string` in `override_values`

This parameter makes the resource use the `--set-string` option of Helm
(instead of the `--set` option) for the given value override. It can be
used to ensure that Helm always treats the value as a string, useful if
the value varies and may look like a number, eg. if it's a Git commit
hash. See https://github.com/kubernetes/helm/issues/1707