feat: Enable pool kubeconfig pattern
Allow the concourse-helm-resource to be used in put-only mode with a
kubeconfig file.

This builds off of pr46, but extends it so that there is no need to
configure the resource with the cluster_url (or anything else specified
in the kubeconfig).

Additionally, add a number of additional related controls:

- Get the k8s namespace from the kubeconfig file
- Ability to configure the tiller namespace
- Ability to enable tracing for debugging the resource
- Expose the helm init --wait option
- Ability to specify the kubeconfig path on put.

Together, these features allow us to use the helm resource in conjunction
with the pool resource and the kubernetes resource to:

- Dynamically pick a k8s namespace from a pool, granting exclusive access
- Run a helm-based standalone integration test against a clean environment
  in a dedicated namespace with a dedicated tiller
  - The concourse-helm-resource cleans up anything that may have been in the
    namespace's tiller
  - The concourse-helm-resource installs helm into the namespace
  - The concourse-helm-resource waits for tiller to be ready
  - The concourse-helm-resource installs the chart into the namespace with
    the namespace-specific tiller
- Gather the results (with the kubernetes resource)
- Return the k8s namespace to the pool